Initial euro_ix_q commit -- work in progress
	modified:   bgpd/Makefile.am
	modified:   bgpd/bgp.h
	new file:   bgpd/bgp_adj_in.c
	new file:   bgpd/bgp_adj_in.h
	new file:   bgpd/bgp_adj_out.c
	new file:   bgpd/bgp_adj_out.h
	modified:   bgpd/bgp_advertise.c
	modified:   bgpd/bgp_advertise.h
	modified:   bgpd/bgp_aspath.c
	modified:   bgpd/bgp_aspath.h
	modified:   bgpd/bgp_attr.c
	modified:   bgpd/bgp_attr.h
	new file:   bgpd/bgp_attr_store.c
	new file:   bgpd/bgp_attr_store.h
	modified:   bgpd/bgp_btoa.c
	modified:   bgpd/bgp_clist.c
	modified:   bgpd/bgp_clist.h
	new file:   bgpd/bgp_cluster.c
	new file:   bgpd/bgp_cluster.h
	modified:   bgpd/bgp_common.c
	modified:   bgpd/bgp_common.h
	modified:   bgpd/bgp_community.c
	modified:   bgpd/bgp_community.h
	modified:   bgpd/bgp_connection.c
	modified:   bgpd/bgp_connection.h
	modified:   bgpd/bgp_damp.c
	modified:   bgpd/bgp_damp.h
	modified:   bgpd/bgp_debug.c
	modified:   bgpd/bgp_debug.h
	modified:   bgpd/bgp_dump.c
	modified:   bgpd/bgp_dump.h
	modified:   bgpd/bgp_ecommunity.c
	modified:   bgpd/bgp_ecommunity.h
	modified:   bgpd/bgp_engine.c
	modified:   bgpd/bgp_engine.h
	modified:   bgpd/bgp_filter.c
	modified:   bgpd/bgp_filter.h
	modified:   bgpd/bgp_fsm.c
	modified:   bgpd/bgp_fsm.h
	modified:   bgpd/bgp_main.c
	modified:   bgpd/bgp_mplsvpn.c
	modified:   bgpd/bgp_mplsvpn.h
	modified:   bgpd/bgp_msg_read.c
	modified:   bgpd/bgp_msg_read.h
	modified:   bgpd/bgp_msg_write.c
	modified:   bgpd/bgp_msg_write.h
	modified:   bgpd/bgp_names.c
	modified:   bgpd/bgp_names.h
	modified:   bgpd/bgp_network.c
	modified:   bgpd/bgp_network.h
	modified:   bgpd/bgp_nexthop.c
	modified:   bgpd/bgp_nexthop.h
	modified:   bgpd/bgp_notification.c
	modified:   bgpd/bgp_notification.h
	modified:   bgpd/bgp_open.c
	modified:   bgpd/bgp_open.h
	modified:   bgpd/bgp_open_state.c
	modified:   bgpd/bgp_open_state.h
	modified:   bgpd/bgp_packet.c
	modified:   bgpd/bgp_packet.h
	modified:   bgpd/bgp_peer.c
	modified:   bgpd/bgp_peer.h
	modified:   bgpd/bgp_peer_index.c
	modified:   bgpd/bgp_peer_index.h
	new file:   bgpd/bgp_peer_vty.c
	new file:   bgpd/bgp_peer_vty.h
	modified:   bgpd/bgp_regex.c
	modified:   bgpd/bgp_regex.h
	new file:   bgpd/bgp_rib.c
	new file:   bgpd/bgp_rib.h
	modified:   bgpd/bgp_route.c
	modified:   bgpd/bgp_route.h
	modified:   bgpd/bgp_route_refresh.c
	modified:   bgpd/bgp_route_refresh.h
	modified:   bgpd/bgp_routemap.c
	modified:   bgpd/bgp_session.c
	modified:   bgpd/bgp_session.h
	modified:   bgpd/bgp_snmp.c
	modified:   bgpd/bgp_table.c
	modified:   bgpd/bgp_table.h
	new file:   bgpd/bgp_unknown.c
	new file:   bgpd/bgp_unknown.h
	modified:   bgpd/bgp_vty.c
	modified:   bgpd/bgp_vty.h
	modified:   bgpd/bgp_zebra.c
	modified:   bgpd/bgp_zebra.h
	modified:   bgpd/bgpd.c
	modified:   bgpd/bgpd.h
	new file:   bgpm/bgpm_south.h
	modified:   isisd/isis_routemap.c
	modified:   lib/Makefile.am
	modified:   lib/avl.c
	modified:   lib/avl.h
	new file:   lib/bitmap.c
	new file:   lib/bitmap.h
	modified:   lib/command.c
	modified:   lib/command.h
	modified:   lib/command_local.h
	modified:   lib/command_parse.c
	modified:   lib/command_parse.h
	modified:   lib/confirm.h
	modified:   lib/elstring.h
	modified:   lib/filter.c
	modified:   lib/filter.h
	modified:   lib/heap.h
	modified:   lib/if.c
	new file:   lib/ihash.c
	new file:   lib/ihash.h
	new file:   lib/iovec.h
	new file:   lib/keyword.c
	new file:   lib/keyword.h
	modified:   lib/list_util.h
	modified:   lib/log.c
	modified:   lib/log.h
	modified:   lib/memory.c
	modified:   lib/memory.h
	new file:   lib/mempool.c
	new file:   lib/mempool.h
	modified:   lib/memtypes.awk
	modified:   lib/memtypes.c
	modified:   lib/misc.c
	modified:   lib/misc.h
	modified:   lib/miyagi.h
	modified:   lib/mqueue.c
	modified:   lib/mqueue.h
	modified:   lib/name_map.c
	modified:   lib/name_map.h
	new file:   lib/pfifo.c
	new file:   lib/pfifo.h
	modified:   lib/plist.c
	modified:   lib/plist.h
	modified:   lib/prefix.c
	modified:   lib/prefix.h
	new file:   lib/prefix_id.c
	new file:   lib/prefix_id.h
	modified:   lib/privs.c
	modified:   lib/pthread_safe.c
	modified:   lib/pthread_safe.h
	modified:   lib/qafi_safi.c
	modified:   lib/qafi_safi.h
	new file:   lib/qatomic.c
	new file:   lib/qatomic.h
	modified:   lib/qfstring.c
	modified:   lib/qfstring.h
	modified:   lib/qiovec.h
	modified:   lib/qlib_init.c
	modified:   lib/qlib_init.h
	new file:   lib/qlump.c
	new file:   lib/qlump.h
	modified:   lib/qpnexus.c
	modified:   lib/qpnexus.h
	modified:   lib/qpselect.c
	modified:   lib/qpselect.h
	modified:   lib/qpthreads.c
	modified:   lib/qpthreads.h
	modified:   lib/qstring.c
	modified:   lib/qstring.h
	modified:   lib/qtime.c
	modified:   lib/qtime.h
	new file:   lib/ring_buffer.c
	new file:   lib/ring_buffer.h
	modified:   lib/routemap.c
	modified:   lib/routemap.h
	modified:   lib/sockopt.c
	modified:   lib/sockopt.h
	modified:   lib/sockunion.c
	modified:   lib/sockunion.h
	modified:   lib/stream.c
	modified:   lib/stream.h
	deleted:    lib/symtab.c
	deleted:    lib/symtab.h
	modified:   lib/table.c
	modified:   lib/vector.c
	modified:   lib/vector.h
	modified:   lib/version.h.in
	new file:   lib/vhash.c
	new file:   lib/vhash.h
	modified:   lib/vio_fifo.c
	modified:   lib/vio_fifo.h
	modified:   lib/vio_lines.c
	modified:   lib/vty.c
	modified:   lib/vty.h
	modified:   lib/vty_io.c
	modified:   lib/vty_io.h
	modified:   lib/vty_io_basic.c
	modified:   lib/vty_io_basic.h
	modified:   lib/vty_io_file.c
	modified:   lib/vty_io_term.c
	modified:   lib/vty_local.h
	modified:   lib/workqueue.c
	modified:   lib/workqueue.h
	modified:   lib/zassert.h
	modified:   lib/zclient.c
	modified:   lib/zclient.h
	modified:   lib/zconfig.h
	modified:   lib/zebra.h
	modified:   ospf6d/ospf6_asbr.c
	modified:   ospfd/ospf_routemap.c
	modified:   ospfd/ospf_zebra.c
	modified:   ripd/rip_routemap.c
	modified:   ripd/rip_zebra.c
	modified:   ripd/ripd.c
	modified:   ripngd/ripng_zebra.c
	modified:   ripngd/ripngd.c
	modified:   tests/Makefile.am
	modified:   tests/aspath_test.c
	modified:   tests/bgp_capability_test.c
	modified:   tests/bgp_mp_attr_test.c
	deleted:    tests/ecommunity_test.c
	modified:   tests/heavy-thread.c
	modified:   tests/main.c
	modified:   tests/test-avl.c
	modified:   tests/test-checksum.c
	modified:   tests/test-heap.c
	modified:   tests/test-list_util.c
	modified:   tests/test-prefix.c
	modified:   tests/test-privs.c
	modified:   tests/test-qfs.c
	new file:   tests/test-qlump.c
	deleted:    tests/test-symtab.c
	modified:   tests/test-vector.c
	new file:   tests/test_bgp_as_path.c
	new file:   tests/test_bgp_attr_store.c
	new file:   tests/test_bgp_cluster.c
	new file:   tests/test_bgp_community.c
	new file:   tests/test_bgp_ecommunity.c
	new file:   tests/test_bgp_unknown.c
	modified:   vtysh/extract.pl.in
	modified:   vtysh/vtysh.c
	modified:   vtysh/vtysh_config.c
	modified:   vtysh/vtysh_main.c
	modified:   vtysh/vtysh_user.c
	modified:   vtysh/vtysh_user.h
	modified:   zebra/zebra_rib.c
	modified:   zebra/zebra_routemap.c
	modified:   zebra/zserv.c

+++ lib/command_local.h
@@ -59,7 +59,7 @@ typedef struct
 /*------------------------------------------------------------------------------
  * Host configuration variable
  */
-typedef unsigned long int name_gen_t ;
+typedef ulong name_gen_t ;      /* accept 32-bits on 32-bit machine !   */
 
 struct host
 {
@@ -110,7 +110,7 @@ struct host
 
   /* Someone has the config symbol of power
    */
-  ulong      config ;           /* zero <=> no owner    */
+  ulong      config ;                   /* zero <=> no owner    */
   ulong      config_brand ;
 
   /* Allow vty to start without password
@@ -123,7 +123,7 @@ struct host
 
   /* vty timeout value -- see "exec-timeout" command
    */
-  unsigned long vty_timeout_val ;
+  uint       vty_timeout_val ;
 
   /* vty access-class for IPv4 and IPv6
    */
